Improved "runningInNodeJS" check
When in a commonJS environment (like webpack), this library encounters
an infinite loop between `newByteArray` and `convertToByteArray`.

Simply hardening the check to see if we are in a node environment
is enough to prevent the loop from happening. Although this can be
spoofed (like many isNode checks), the assumption I am working
off of is checking the existance of `process.versions.node`.

This won't cause new problems, but also isn't bullet proof either.
And its hopefully better than `typeof window === "undefined"` :)